% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiguresAndTables.R
\name{createFiguresAndTables}
\alias{createFiguresAndTables}
\title{Create figures and tables for the paper}
\usage{
createFiguresAndTables(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema, cdmVersion, folder)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the \code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include
both the database and schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.}

\item{cdmVersion}{Version of the CDM. Can be "4" or "5"}

\item{folder}{(Optional) Name of local file to place results; make sure to use forward slashes (/)}
}
\value{
Creates CSV and PNG files in the specified folder.
}
\details{
This function creates the figures and tables specified in the protocol, based on the data files generated using the \code{\link{execute}} function.

Note that this function requires access to a CDM database to query the vocabulary.
}
