% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Drug Utilization in Children}
\usage{
execute(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema, cdmVersion = 4,
  folder = "DrugsInPeds", file = "StudyResults.zip.enc")
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the \code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include
both the database and schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.}

\item{cdmVersion}{Version of the CDM. Can be "4" or "5"}

\item{folder}{(Optional) Name of local file to place results; make sure to use forward slashes (/)}
}
\value{
Study results are placed in CSV format files in specified local folder. The CSV files are then compressed and encrypted into
a single file called StudyResults.zip.enc.
}
\details{
This function executes the OHDSI Drug Utilization in Children Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        cdmVersion = "4")

}

}

